* dfp.c (encode_decimal32): Type pun via uint32_t. 	(encode_decimal64, encode_decimal128): Likewise.
	Not to be merged:
	* gdbinit.in: Don't set complaints.


git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/dfp-branch@102898 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/dfp.c
@@ -143,7 +143,7 @@ encode_decimal32 (const struct real_format *fmt ATTRIBUTE_UNUSED,
   decimal_to_decnumber (r, &dn); 
   decimal32FromNumber(&d32, &dn, &set);
 
-  buf[0] = *(long *)d32.bytes;
+  buf[0] = *(uint32_t *) d32.bytes;
 }
 
 void decode_decimal32 (const struct real_format *fmt ATTRIBUTE_UNUSED,
@@ -168,13 +168,13 @@ encode_decimal64 (const struct real_format *fmt ATTRIBUTE_UNUSED,
 
   if (FLOAT_WORDS_BIG_ENDIAN)
     {
-      buf[0] = *(long *)&d64.bytes[0];
-      buf[1] = *(long *)&d64.bytes[4];
+      buf[0] = *(uint32_t *) &d64.bytes[0];
+      buf[1] = *(uint32_t *) &d64.bytes[4];
     }
   else
     {
-      buf[1] = *(long *)&d64.bytes[0];
-      buf[0] = *(long *)&d64.bytes[4];
+      buf[1] = *(uint32_t *) &d64.bytes[0];
+      buf[0] = *(uint32_t *) &d64.bytes[4];
     }
 }
 
@@ -200,17 +200,17 @@ encode_decimal128 (const struct real_format *fmt ATTRIBUTE_UNUSED,
 
   if (FLOAT_WORDS_BIG_ENDIAN)
     {
-      buf[0] = *(long *)&d128.bytes[0];
-      buf[1] = *(long *)&d128.bytes[4];
-      buf[2] = *(long *)&d128.bytes[8];
-      buf[3] = *(long *)&d128.bytes[12];
+      buf[0] = *(uint32_t *) &d128.bytes[0];
+      buf[1] = *(uint32_t *) &d128.bytes[4];
+      buf[2] = *(uint32_t *) &d128.bytes[8];
+      buf[3] = *(uint32_t *) &d128.bytes[12];
     }
   else
     {
-      buf[0] = *(long *)&d128.bytes[12];
-      buf[1] = *(long *)&d128.bytes[8];
-      buf[2] = *(long *)&d128.bytes[4];
-      buf[3] = *(long *)&d128.bytes[0];
+      buf[0] = *(uint32_t *) &d128.bytes[12];
+      buf[1] = *(uint32_t *) &d128.bytes[8];
+      buf[2] = *(uint32_t *) &d128.bytes[4];
+      buf[3] = *(uint32_t *) &d128.bytes[0];
     }
 }
